+++ src/library/Identicon/index.tsx
@@ -1,27 +1,79 @@
 // Copyright 2023 @paritytech/polkadot-staking-dashboard authors & contributors
 // SPDX-License-Identifier: Apache-2.0
 
-import { Identicon as IdenticonDefault } from '@polkadot/react-identicon';
-import { useApi } from 'contexts/Api';
+import classNames from 'classnames';
+import { useCallback } from 'react';
 import styled from 'styled-components';
-import type { IdenticonProps } from './types';
+import type { Circle } from './circles';
+import { getCircleXY, OUTER_CIRCLE, renderCircle, Z } from './circles';
+import { getColors } from './colors';
+
+export interface IdenticonProps {
+  size: number;
+  disableCursorCopy?: boolean;
+  disableClipboardCopy?: boolean;
+  colors?: string[];
+  value: string;
+}
 
 const Wrapper = styled.div`
   svg > circle:first-child {
     fill: var(--border-primary-color);
   }
 `;
-export const Identicon = ({ value, size }: IdenticonProps) => {
-  const { ss58 } = useApi().network;
+export const Identicon = ({
+  size,
+  value,
+  disableCursorCopy = false,
+  disableClipboardCopy = false,
+  colors: initialColors,
+}: IdenticonProps) => {
+  const xy = getCircleXY();
+
+  const defaultColors =
+    initialColors || new Array<string>(xy.length).fill('#ddd');
+
+  const colors = value ? initialColors || getColors(value) : defaultColors;
+
+  const copyToClipboard = useCallback(() => {
+    if (disableClipboardCopy) {
+      return;
+    }
+    if (navigator) {
+      navigator.clipboard.writeText(value);
+    }
+  }, [value]);
+
   return (
     <Wrapper>
-      <IdenticonDefault
-        value={value}
-        size={size}
-        theme="polkadot"
-        style={{ cursor: 'default' }}
-        prefix={ss58}
-      />
+      <div
+        aria-hidden="true"
+        onClick={() => copyToClipboard}
+        className={classNames({
+          'cursor-copy': !(disableCursorCopy || disableClipboardCopy),
+        })}
+      >
+        <svg
+          height={size}
+          id={value}
+          name={value}
+          viewBox="0 0 64 64"
+          width={size}
+        >
+          {[OUTER_CIRCLE]
+            .concat(
+              xy.map(
+                ([cx, cy], index): Circle => ({
+                  cx,
+                  cy,
+                  fill: colors[index],
+                  r: Z,
+                })
+              )
+            )
+            .map(renderCircle)}
+        </svg>
+      </div>
     </Wrapper>
   );
 };
