--- src/library/Identicon/types.ts
@@ -1,7 +1,0 @@
-// Copyright 2023 @paritytech/polkadot-staking-dashboard authors & contributors
-// SPDX-License-Identifier: Apache-2.0
-
-export interface IdenticonProps {
-  value: string;
-  size?: number;
-}
