+++ src/pages/Overview/index.tsx
@@ -63,7 +63,6 @@ export const Overview = () => {
   return (
     <>
       <PageTitle title={`${t('overview.overview')}`} />
-
       <PageRow>
         <PageHeading>
           <ActiveAccounts />
